Alway skip failing UseSpa tests on Alpine
